refactored files to use dynamodb-enhanced dependency
