+++ src/main/java/com/revature/get_user/models/Tag.java
@@ -1,0 +1,11 @@
+package com.revature.get_user.models;
+
+import lombok.Data;
+import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
+
+@Data
+@DynamoDbBean
+public class Tag {
+    private String name;
+    private String color;
+}
